refactor script name in tests/worklfows to account for pytest convention adjust ci.yml to not ignore Workflow folder
