+++ .github/workflows/ci.yml
@@ -28,4 +28,4 @@ jobs:
       - name: Unittests
         run: |
           venv\Scripts\Activate.ps1
-          pytest -v tests --ignore=tests/workflows
+          pytest -v tests
