+++ tests/test_01_material_application.py
@@ -24,7 +24,7 @@ class TestMaterialAPI:
         self.results_file = os.path.join(self.local_path, "workflows", "test01_results.json")
         reference_file = os.path.join(self.local_path, "workflows", "test_01_reference_results.json")
         self.clean_results(self)  # no idea why but you have to pass there self
-        scdm_script_path = os.path.join(self.local_path, "workflows", "test_01_run_material_application.py")
+        scdm_script_path = os.path.join(self.local_path, "workflows", "run_test_01_material_application.py")
         run_scdm_batch(SCDM_VERSION, API_VERSION, scdm_script_path)
 
         with open(self.results_file) as file:
