+++ tests/workflows/run_test_03_speos_sensors.py
